WIP: debugging race condition.  isolated to buffer's append, lock is needed for the sequence of isBlocked, ptr fetch_add, and branching based on fetch_add.  also observing that pointers are incremented but values not copied (same number of missing and extra (0) elements.

+++ src/io/test/check_buffer.cpp
@@ -81,8 +81,8 @@ int main(int argc, char** argv) {
   size_t bufferCapBefore = tlBuffer.getCapacity();
   size_t bufferCapAfter = 0;
 
-  size_t bufferSizeBefore = tlBuffer.getApproximateSize();
-  size_t bufferSizeAfter = 0;
+  int bufferSizeBefore = tlBuffer.getApproximateSize();
+  int bufferSizeAfter = 0;
 
   const void* bufferPtrBefore = tlBuffer.getData();
   const void* bufferPtrAfter = nullptr;
